+++ Runtime/Helpers/Debug/GizmoHelper.cs
@@ -1,6 +1,6 @@
 using UnityEngine;
 
-namespace LiteNinja.Utils
+namespace LiteNinja.Utils.Helpers
 {
     public static class GizmoHelper
     {